Added UAC prompt
Added the ability to detect admin status and promt the user for elevation if necessary. code shamelessly copied from stackoverflow.

was doing other things and decided to use this script, got annoyed when it reee'd about access denied, so I went on a slight side-quest to change this this. subject to change